defining a deserialise package structure along with model interfaces
